refactor: fix db issues and restructure arch
The app had issues connecting to a database online because well, there
is no cloud database. I took me a while but I finally a platform for
hosting my sql db. I also fix the env issues and the restructured the
application for scalability.

Now, I will move over to the business logic of the app.